refactor: move deprecated things into deprecated dir
