+++ packages/react-dataparcels/ParcelBoundaryHoc.js
@@ -1,2 +1,2 @@
 /* eslint-disable */
-module.exports = require('./lib/ParcelBoundaryHoc.js');
+module.exports = require('./lib/deprecated/ParcelBoundaryHoc.js');
